executor: move some test cases
