+++ pkg/executor/BUILD.bazel
@@ -453,7 +453,6 @@ go_test(
         "//pkg/sessionctx/variable",
         "//pkg/sessiontxn",
         "//pkg/statistics",
-        "//pkg/statistics/handle/ddl/testutil",
         "//pkg/statistics/handle/storage",
         "//pkg/statistics/util",
         "//pkg/store/copr",
@@ -526,7 +525,6 @@ go_test(
         "@org_golang_google_grpc//:grpc",
         "@org_uber_go_atomic//:atomic",
         "@org_uber_go_goleak//:goleak",
-        "@org_uber_go_zap//:zap",
         "@org_uber_go_zap//zapcore",
     ],
 )
